Simplify requests usage
Verbs are preferably upper case.

Add join_path function to join path components at each level in the
class hierarchy.

Modify verb methods (_get, etc.) and _requests and all methods in the
respective call chain to have an optional uri parameter. If the uri is
given, then it's a relative uri that is joined with the instance_url and
base_uri, if any. If an url is given, then that is used as the complete
url. Also update these methods to pass along the remaining kwargs up
the call chain up to requests.
Format the complete url, if not given explicitly, in the Connection
class, which is used by all _Service classes and since the properties
used belong to the Connection instance.

No need to set an Accept header, as Salesforce seems to not rely on that
and requests by default will set it to '*/*'. No need to explicitly set
the Content-Type header, since requests will set that appropriately when
using the json parameter. Update code to use the json parameter
accordingly.

A remaining issue not fixed here is that the _setup_session methods in
RestService and AsyncService both modify a common connection.session
instance, setting headers globally that will be sent in all requests
done from all _Service instances (as they all share the same
connection.session). I'd think the intention was to set the headers only
for requests done in instances of each of the mentioned _Service
classes, respectively. OOP is tricky.

Optimize imports and add missing imports from previous commit.